Distinguish infra errors from kills; align u32; avoid octal literals
Three correctness bugs collapsed into one round of fixes.

(1) runner exits 124 (timeout), 125 (wait fail), 127 (spawn fail), and
2 (command unparseable) were all routed through 'if exit_code != 0 ->
continue mutate', so a typo in --command turned every spec into a
silent kill. Reports claimed 0 failures with confidence. runner now
exports INFRA_EXIT_CODES + is_infra_error; main warns and continues
to the next spec without crediting a kill, and AnalyticsResult::failures
mirrors the same rule so the headline counter agrees with the loop.

(2) extract_glob_pattern parsed the line-number suffix as u16 while
extract_line_number used u32 (commit 29b5595 widened one and missed
the other). A line number above 65535 silently dropped to 'no match' —
no error, no warning, an empty mutation set on huge generated files.
Single-character widen.

(3) generate(3, '1234567890') produced ~10% leading-zero strings.
Ruby parses 010 as octal 8 and 088 as a syntax error; the latter
becomes another infra failure mode now handled by (1), the former
silently mutates to a different decimal than the digits suggest.
Split the 3-digit replacement into one leading non-zero digit plus
two free digits.

+++ engine/src/file/mod.rs
@@ -58,7 +58,7 @@ impl File {
 
     pub fn extract_glob_pattern(path: &str) -> &str {
         match path.rsplit_once(':') {
-            Some((prefix, tail)) if tail.parse::<u16>().is_ok() => prefix,
+            Some((prefix, tail)) if tail.parse::<u32>().is_ok() => prefix,
             _ => path,
         }
     }
@@ -70,10 +70,7 @@ impl File {
     }
 
     pub fn find_mutations(file_path: String, line_number: u32) -> Vec<MutableItem> {
-        match Path::new(&file_path)
-            .extension()
-            .and_then(|s| s.to_str())
-        {
+        match Path::new(&file_path).extension().and_then(|s| s.to_str()) {
             Some("rb") => ruby::find_all(&file_path, line_number),
             _ => {
                 warn!("File '{}' is not supported. Skipping.", file_path);
